feat: rename discussion folder -> discussions
