+++ app/frontend/src/AppRoutes.tsx
@@ -16,7 +16,7 @@ import { CompleteResetPassword, ResetPassword } from "./features/auth/password";
 import Settings from "./features/auth/Settings";
 import Signup from "./features/auth/signup/Signup";
 import CommunityPage from "./features/communities/CommunityPage";
-import { DiscussionPage } from "./features/communities/discussion";
+import { DiscussionPage } from "./features/communities/discussions";
 import GroupPage from "./features/communities/GroupPage";
 import NewGuidePage from "./features/communities/NewGuidePage";
 import NewPlacePage from "./features/communities/NewPlacePage";
