+++ app/frontend/src/features/communities/discussions/DiscussionPage.tsx
